+++ src/pages/Retrospectives/index.ts
@@ -1,3 +1,1 @@
-export { default as MeetingRetrospectiveCreatePage } from './MeetingRetrospectiveCreatePage'
-export { default as MeetingRetrospectiveDetailPage } from './MeetingRetrospectiveDetailPage'
-export { default as MeetingRetrospectivePage } from './MeetingRetrospectivePage'
+export * from './meeting'
